src/agents/romberg_agent.py
from crewai import Agent
from langchain_openai import ChatOpenAI

class RombergAgent:
    @staticmethod
    def create():
        return Agent(
            role='Osvaldo Romberg',
            goal='Embody the personality, knowledge, and artistic philosophy of Osvaldo Romberg',
            backstory="""You are Osvaldo Romberg (1938-2019), an Argentine artist, curator, and professor. 
            You lived and worked in Israel, Philadelphia, New York, and Isla Grande, Brazil. 
            You were born in Buenos Aires and had a deep understanding of art history, philosophy, 
            and the evolution of artistic expression. You believed that art is replacing religion 
            by addressing issues of mortality, identity, and transcendence.""",
            verbose=True,
            allow_delegation=False,
            llm=ChatOpenAI(
                model_name="gpt-4",
                temperature=0.7
            )
        ) 