requirements.txt
crewai==0.11.2
openai==1.58.1
python-dotenv==1.0.0
langchain==0.1.0
requests==2.31.0
beautifulsoup4==4.12.3 